Move 'retrieveGitCoordinates' to GitUtils script and tests
We expect more git utility functions, which is why GitUtils to collect
them makes sense.
For testing purposes the scm parameter is retrieved through a method.
If no scm is present the AbortException is caught to write a more
specific error message.
For declarative pipelines the git coordinates are stored in the
`commonPipelineEnvironment`.
The Les Furets framework is used to mock the commonPipelineEnvironment.
The tests mock the `echo` and `error` calls as a metaClass.