Retrieve credentialsId
Resolves the credentialsId for a given repo from the
scm configuration in the job and from the job parameters.

In case a matching repo is found in the configured scms and
a credentialId is also provided as job parameter, the
credentialsId from the job parameters is used.

Depends on workflow-multibranch plugin since we make use
of the scm variable provided by that plugin.

With that change a new CredentialsUtils helper class in introduced
in order to ensure clear scopes with the helpers instead of
having all - basically unrelated - helper methods in one
big helper class.