Remove Utils#retrieveGitCoordinates
The methods a related to an outdated use case

It was intended for getting the coordinates of a payload repo when the
pipeline repo is specified in the pipeline job.

That has been changed so that the pipeline (or at least a stub pipeline)
is now always contained in the payload repo and the payload repo is
always specified in the pipeline job itself.