Change parameter name 'changeId' to 'changeDocumentId'
the official term in SAP-SolutionManager is 'change document'.

The term 'change' as reflected in 'changeId' is ambigous and could also be understand in the
sense of a change in a revision control system.

Choosing appropriate terms for parameters here deceases the amount of documentation
required for explaining what is really denoted by the parameter.