piperStageWrapper - add script as parameter
Based on interactions with teams we saw that the current extension mechanism is insufficient.

When calling other library steps we need to have the current `script` available. One reason is to be able to properly pass the context. This is for example required to resolve step configuration. Currently, the config parameter only contains the general as well as the relevant stage section of the configuration since at this point no information is available about which steps are executed in the extension.

This PR aims to add the functionality of having the main script context inside an extension.

An alternative solution would be to change the extension syntax to only pass a Map instead of multiple parameters.
This way we would be save with potential future adaptions and can stay compatible.

Personally, I opt for choosing a map. If this was common agreement I would adapt this PR accordingly.