dockerExecute - extend Kubernetes capabilities
allow `dockerExecute` to pass on to dockerExecuteOnKubernetes
* containerCommand
* containerShell

`containerCommand` is required in case a container comes with a shell where the default command to keep the container running is not applicable, e.g. when using a busybox shell

`containerShell` defines the shell to use in a Kubernetes scenario. In this scenario it is not sufficient to pass the shell via shebang. It needs to passed explicitly to the exeuction of the pod/container.

These extensions enable users to use dockerExecute (instead of dockerExecuteOnKubernetes) and stay independent of the underlying infrastructure by passing all relevant parameters to `dockerExecute` and avoid a conditional in the calling step to decide on the infrastructure.