Retrieve used steps from test run
We would like to have a where-use-list which contains the
directly required plugins for a step.

With this change we keep track which steps are called from a pipeline
step test. This can be used an input for making queries on a Jenkins
which plugins provides the steps in question.

Better approach would be to hook into PipelineTestHelper#methodInterceptor
(... a closure). But this is not that easy and directly modiyfing the
code in the JenkinsPipelineTesting framework is IMO also not a cool
option.