take parameters keys into account again.
was removed since parameter keys does not appear anymore in the table denoting the
level where a parameter can be configured.

Nevertheless we need to take that parameter into account. We need to add the description for
that parameter, even if it cannot be configured on general or stage level.