Resolve calls to plugin steps in a transitive way
We have the case that one piper step calls other piper steps.
In such cases we would like to get all the plugin calls. The direct
calls but also the calls performed by nested piper-lib step calls.

The plugin calls as we registed them during listing to the tests are
resolved with the script provided by this commit.