Provide prelininary log files in case test takes unusually long
In some case Jenkins cannot be launched, but does also not die (remains
in an undefinded status). In this case the integration tests never comes
back and the corresponding job gets killed by the build intfastructure.

In this case we provide the logs written until this point in time in order
to have some hints for resolving the issue.

Currently the integration tests gets killed after 50 min. The preliminary
logs are written after 40 min. Normally intgration tests should never take
that long.