fetch possible values from @see
In case there is a @see tag provided for a parameter in the description,
the possible values are only fetched from the corresponding source in
case there is no explict value provided for the possible values.

In case the source does not contain a possible value tag nothing is
transfered.