Check against containerMap from config rather than params (#835)
The parameter map is directly handed over from outside into the step via signature of the call method.

The container map is defined as step parameters, not as parameter handed over (only) via the parameters map.

With the current approach only the container map from the parameters is taken into account. In case the
parametersMap is defined elsewhere it is not taken into account.