Emit piper go version to the log after materialization
Might be useful when troubleshooting.

Would be better to compare the git hash with the git hash of the current version of the piper lib.
But in order to get that we have to retrieve that git hash for the current version of the piper lib.
That is not that easy. For the time being that is better than nothing.