Store instance as property to script binding
For us not to lose the instance when the pipeline is stopped we store
its values away in the script. If there is no instance but the
respective property is set, we create an instance with the values that
were stored away previously.
This means we need to give the script binding to the DefaultValueCache
class to store the default values from the instance in its binding.