Set revision in MavenArtifactVersioning
In case of maven ci-friendly versioning (https://maven.apache.org/maven-ci-friendly.html)
a revision property is used to define the version of the project.
When we set the version of the project we also have to update the revision
because there could be in-project dependencies (e.g. integration-tests)
which refer to the test artifacts version as ${revision}.