artifactPrepareVersion: new versioning option
add versioning option `cloud_noTag`.
This allows teams to use automatic versioning
without creating a tag on the SCM which contains the change.

This option needs to be handled with care since
teams can create a build artifact with no real representation in the SCM.

Thus the `cloud_noTag` option is weak when it comes to
providing a tracable build result