Enable use of customDefaults from Project Config
This change enables the setupCommonPipelineEnvironment step to handle
custom default configurations defined in customDefaults parameter of the
project configuration.

Previously, only the getConfig Go step was able to incorporate custom
default configurations provided in the customDefaults parameter of the
project configuration. Now, also the setupCommonPipelineEnvironment and
DefaultValueCache provide the functionality the getConfig Go step
already provides.