Add stashing to piperExecuteBin
An optional List is passed where each entry is a Map describing parameters for stashing and un-stashing.
When executing on docker, it is assumed that the files need to be un-stashed first.
Directly after executing piper-go, the files are stashed (i.e. either while still in docker, or likewise when not).