Detect an error response
Receiving an error response means receiving a response body succeeded. It is valid JSON, so unmarshalling it also works, even if not a single field could be matched into the provided Go structure. Try to unmarshall into a dedicated error response structure and check if the response contained an "errorCode" field.