Add a milestone for Confirm stage
This commit will prevent abortion of older builds which are waiting in the Confirm stage,
if a newer build fails in the last stage before the Confirm stage.

Example: If the 'Compliance' stage of the general purpose pipeline (piperPipeline.groovy) fails,
it will abort all former builds which are waiting in the 'Confirm' stage.

The stageLocking nature of piperStageWrapper is disabled for the Confirm stage as we still
want to be able to promote any of the builds waiting for the manual confirmation.

Note: Confirming a build will still abort all older builds waiting at the Confirm stage.