Avoid test failure caused by inconsistencies wrt quotation marks
When executing the tests I get

=== 8< ===
    --- FAIL: TestRunGitopsUpdateDeploymentWithHelm/erroneous_URL (0.00s)
        gitopsUpdateDeployment_test.go:392:
                Error Trace:    gitopsUpdateDeployment_test.go:392
                Error:          Error message not equal:
                                expected: "failed to apply helm command: failed to extract registry URL, image name, and image tag: registry URL could not be extracted: invalid registry url: parse \"://myregistry.com\": missing protocol scheme"
                                actual  : "failed to apply helm command: failed to extract registry URL, image name, and image tag: registry URL could not be extracted: invalid registry url: parse ://myregistry.com: missing protocol scheme"
                Test:           TestRunGitopsUpdateDeploymentWithHelm/erroneous_URL
=== 8< ===

Since this does not happen during the central build it might be
mac specific.

The reason for the failure are the quotation marks surrounding
the (invalid) url.

In order to avoid this incompatibility we assert the error message
of the innermost error in the chain of errors.

Nota bene: the (currently) latest release of testify does not provide
any helper for making assertion on errors contained in an error chain.
Starting with the next release there should be ErrorIs. We can switch
to this at a later point in time