Reduce severity in case of http 401, 403, 404 and 500
It is not necessarily an error. E.g in case of a webanalytics call resulting in a 404 we
should not have a message with severity 'error' in the log since the 'payload' task itself
was successful.

Maybe it would even better not to log something here, but only let the error bubble up and
let the caller decide.