move additional hints into markdown file
These hints are up to now provided as javadoc like comments
at the call method for cloudFoundryDeploy step. At that level
they are not taken into account by the go generator.

Is is not a suitable approach to include this kind of markdown
docu into the long description of the metadata file since this
long description is (also) used by the piper help command when
invoking piper cloudFoundyDeploy --help. At that point in time
markdown like text does not make sense.