Introduce useGoStep parameter in go metadata
this parameter has no effect in go, but it is used in the groovy layer. When we
publish the documentation already based on the go layer this parameter would not
be mentioned as long as the groovy step is not purged.

This parameter has to be removed after purging the groovy layer (... when there is only the
normal stub contained in groovy).