Measure the duration also in the case of failures
So far the duration is not stored in the InfluxData, if the measured build step throws an Exception (e.g. because of an timeout). So Grafana (or similar tools) shows only the build time of successful build steps.
This change catches the potential Exception or Error, stores the duration and re-throws the Exception or Error after storing.