Allow empty default stashes
This makes it so that for the default stash/unstash (before and after build)
allowEmpty is set to true therefore not generating an Error. The corresponding
Exception is caught however generating an Error leads to the Error getting
recorded in Jenkins. Since Piper defaults - which can not be easily changed -
include a minimum set of stashes to be created it leads to more user happiness
if no messages are created.