fix(JenkinsConfigProvider): panic in GetBuildReason
the `_class` field of the objects in the actions array is not always
populated, like in this example (notice the `CauseAction` is only the
2nd entry):

```json
{
  "_class": "org.jenkinsci.plugins.workflow.job.WorkflowRun",
  "actions": [
    {},
    {
      "_class": "hudson.model.CauseAction",
      "causes": [
        {
          "_class": "hudson.model.Cause$UserIdCause",
          "shortDescription": "Started by user John Doe",
          "userId": "d012345",
          "userName": "John Doe"
        },
        {
          "_class": "org.jenkinsci.plugins.workflow.cps.replay.ReplayCause",
          "shortDescription": "Replayed #1"
        }
      ]
    },
    {
      "_class": "jenkins.scm.api.SCMRevisionAction"
    }
  ]
}
```

I have no idea why this happens, but it causes the `GetBuildReason`
method to panic. This PR fixes this by ignoring any actions without
class.

Co-authored-by: Philipp Stehle <philipp.stehle@sap.com>